+++ CHANGELOG.md
@@ -3,6 +3,9 @@ All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
+### Added
+- Various example apps
+
 ### Fixed
 - `npm test` command not working.
 - Removed *.spec files from coverage test.