+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ### Fixed
 - `npm test` command not working.
 - Removed *.spec files from coverage test.
+- Already parsed values are not parsed again.
 
 ## [0.4.0] - 2016-04-22
 ### Added